project-c-video-transcription/requirements.txt
# Video upload + transcription API
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9
sqlalchemy>=2.0.0
openai>=1.0.0
